# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Takomo
  version: "1.0.0-draft1"
  description: |
    Central task store for agent fleets. One authority, bearer-token auth,
    hierarchical tickets, enforced per-project workflows, atomic claim/lease,
    append-only event log.

    Design rules this contract encodes:
    - Most writes are commutative (comments, field sets, metadata merge); CAS
      (`If-Match` on `version`) is only needed for whole-body replacement.
    - Every rejection teaches: error bodies state what is allowed and how to
      proceed (`allowed_transitions`, `remedy`).
    - Claiming is one atomic operation (`POST /ready/claim`), never read-then-claim.
    - The event log (`global_seq`) is the only push/poll surface; all delivery
      mechanisms (SSE, hooks, channels) layer on the `since` cursor.
    - `/v1` is frozen: additive evolution only; experiments live in `metadata`.
servers:
  - url: /v1
security:
  - bearer: []
...
paths:
...
  /tickets/{id}:
    get:
      tags: [tickets]
      summary: Get a ticket
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: include, in: query, schema: { type: string }, description: "comma list: comments,children,deps,events" }
      responses:
        "200":
          description: OK
          headers:
            ETag: { schema: { type: string }, description: current version, e.g. "7" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [tickets]
      summary: Update fields (commutative; CAS only for body)
      description: |
        Field-level sets (`title`, `priority`, `labels`) and `metadata_merge`
        (RFC 7386 merge-patch semantics, namespaced keys) are commutative and
        never require `If-Match`. `links` MERGES per key: each key in the patch
        is set (string value) or deleted (null value) while untouched keys
        survive, so concurrent writers attaching different link keys never
        clobber each other. Replacing `body` requires `If-Match: "<version>"`;
        on mismatch the 409 carries the current version and body hash. State is
        NOT patchable — use /transition.
        While a ticket is claimed, non-holders may only patch `metadata_merge`
        under their own namespace and add comments; other fields 409.
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: If-Match, in: header, required: false, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "409": { $ref: "#/components/responses/Conflict" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tickets/{id}/transition:
    post:
      tags: [workflow]
      summary: Move a ticket to another state (the only way state changes)
      description: |
        Validated against the project workflow: transition must exist from the
        current state, all `requires` must hold (active claim held by caller,
        token scope, guards like no-open-children). Illegal transitions return
        409 with `allowed_transitions` and a `remedy` — errors are written for
        LLM readers.

        Checks run in the order legality → scope → claim/fence, and the headline
        `code`/`message` names the first real blocker in that order: an illegal
        target is `transition.illegal`, a missing scope is `transition.scope`
        (403), and only then are claim ownership and fence echoing checked. So a
        lease holder attempting a human gate they lack the scope for sees the
        403, not a fencing error.

        Human approval is authoritative over a lease: a transition whose
        `requires` includes `scope:human`, performed by a caller holding the
        `human` scope, succeeds even while another actor holds the claim, and
        auto-releases that claim as a side effect (emitting a `released` event
        alongside `transitioned`). This override is scoped to human-required
        transitions; ordinary `claim`-required transitions keep the holder lock.
      parameters: [ { $ref: "#/components/parameters/id" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, description: target state id }
                reason: { type: string, maxLength: 2000 }
                fence: { type: integer, description: required while ticket is claimed }
      responses:
        "200":
          description: Transitioned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "409":
          description: Illegal transition, missing claim, stale fence, or guard failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitionError" }
        "403":
          description: Transition requires a scope this token lacks (e.g. human)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitionError" }
...
  /shares/self/tickets/{id}:
    get:
      tags: [shares]
      summary: One in-scope ticket with comments + deps — share token
      security:
        - shareBearer: []
      description: |
        Authenticated by the SHARE token. Returns one ticket (with `comments`
        and `deps`) for the board's detail panel. A ticket outside the share's
        scope returns `404` — the share cannot reach it.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410":
          description: The shared link has expired or been revoked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: See auth.md — `tk_`-prefixed tokens, scoped, hashed at rest.
    shareBearer:
      type: http
      scheme: bearer
      description: |
        A SHARE token (`tks_`-prefixed), minted by `POST /v1/shares` and hashed
        at rest like a normal token. It is a distinct auth path: it authorizes
        ONLY the read-only `/v1/shares/self*` endpoints, bounded to its scope,
        and is rejected on every normal endpoint. Expired/revoked shares return
        `410 Gone`. See auth.md.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Ticket id, e.g. "rvp-x7k2"
...
  responses:
    NotFound:
      description: Unknown id
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Version mismatch, claim held by another, or stale fence
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Ticket:
      type: object
      properties:
        id: { type: string }
        project: { type: string }
        type: { type: string, enum: [epic, task, bug, spike] }
        parent: { type: [string, "null"] }
        title: { type: string, maxLength: 300 }
        body: { type: string, maxLength: 131072, description: markdown }
        state: { type: string, description: workflow state id }
        state_category: { type: string, enum: [todo, in_progress, blocked, review, done, cancelled] }
        priority: { type: string, enum: [critical, high, normal, low], default: normal }
        labels: { type: array, items: { type: string } }
        metadata:
          type: object
          description: freeform JSON; keys namespaced "<actor-or-tool>.<key>"; 64 KiB cap
        links:
          type: object
          description: well-known evidence pointers
          properties:
            branch: { type: string }
            pr: { type: string, format: uri }
            env: { type: string, format: uri }
            run: { type: string }
        blocked_by: { type: array, items: { type: string } }
        claim:
          type: [object, "null"]
          properties:
            holder: { type: string, description: actor name from token }
            expires_at: { type: string, format: date-time }
        version: { type: integer, description: monotonic per ticket; ETag value }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: [string, "null"], format: date-time, description: when archived (hidden from default views), or null when active }
...
    TicketPatch:
      type: object
      description: All fields optional; state is absent by design (use /transition)
      properties:
        title: { type: string }
        body: { type: string, description: requires If-Match }
        priority: { type: string }
        labels: { type: array, items: { type: string } }
        labels_add: { type: array, items: { type: string } }
        labels_remove: { type: array, items: { type: string } }
        parent: { type: [string, "null"] }
        links: { type: object, description: "per-key merge: string value sets a key, null deletes it, untouched keys survive" }
        metadata_merge: { type: object, description: RFC 7386 merge-patch; null deletes a key }
        fence: { type: integer, description: required while ticket is claimed and caller is holder }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "stable machine code, e.g. conflict.version, claim.held, fence.stale (superseded lease), fence.invalid (fence never issued — a client bug)" }
        message: { type: string, description: written for LLM readers — states what is wrong AND what to do }
        remedy: { type: string, description: the exact next call to make, when one exists }
        current_version: { type: integer }
        details: { type: object }
    TransitionError:
      allOf:
        - $ref: "#/components/schemas/Error"
      type: object
      properties:
        current_state: { type: string }
        allowed_transitions:
          type: array
          items:
            type: object
            properties:
              to: { type: string }
              requires: { type: array, items: { type: string } }